+++ src/simmate/command_line/base_command.py
@@ -51,21 +51,31 @@ def run_server():
 
     This command is exactly the same as running:
 
-    `django runserver --settings=simmate.configuration.django.settings`
+    `django-admin runserver --settings=simmate.configuration.django.settings`
     """
 
-    # BUG: logging prints ugly duplicates as it sets up simmate and also a
+    # BUG: windows dev version is throwing issues with this code section here,
+    # so I just switch to using subprocess below.
+    # ---> Error while finding module specification for '__main__'
+    # (ValueError: __main__.__spec__ is None)
+    #
+    # logging prints ugly duplicates as it sets up simmate and also a
     # static server for it.
-    logging.warning(
-        "Seeing duplicate messages is normal and expected. "
-        "This is because the test server sets up + tears down Simmate on a cycle."
+    # logging.warning(
+    #     "Seeing duplicate messages is normal and expected. "
+    #     "This is because the test server sets up + tears down Simmate on a cycle."
+    # )
+    # from django.core.management import call_command
+    # from simmate.database import connect
+    # call_command("runserver")
+
+    import subprocess
+
+    logging.info("Setting up local test server...")
+    subprocess.run(
+        "django-admin runserver --settings=simmate.configuration.django.settings",
+        shell=True,
     )
-
-    from django.core.management import call_command
-
-    from simmate.database import connect
-
-    call_command("runserver")
 
 
 @simmate_app.command()