+++ src/simmate/database/base_data_types/nudged_elastic_band.py
@@ -20,6 +20,14 @@ class DiffusionAnalysis(Structure):
     class Meta:
         app_label = "workflows"
 
+    api_filters = dict(
+        migrating_specie=["exact"],
+        vacancy_mode=["exact"],
+        atomic_fraction=["range"],
+        barrier_cell=["range"],
+        npaths_involved=["range"],
+    )
+
     migrating_specie = table_column.CharField(max_length=4, blank=True, null=True)
     """
     The element of the diffusion atom (e.g. "Li")
@@ -146,6 +154,16 @@ class MigrationHop(DatabaseTable):
         "index_end",
         "number",
     ]
+
+    api_filters = dict(
+        site_start=["exact"],
+        site_end=["exact"],
+        number=["range"],
+        length=["range"],
+        dimension_path=["range"],
+        dimension_host_lattice=["range"],
+        energy_barrier=["range"],
+    )
 
     # OPTIMIZE: site_start and site_end
     # Really, this is a list of float values, but I save it as a string.
@@ -396,6 +414,13 @@ class MigrationImage(Structure):
         "structure_distance",
         "energy",
     ]
+
+    api_filters = dict(
+        number=["exact"],
+        force_tangent=["range"],
+        energy=["range"],
+        structure_distance=["range"],
+    )
 
     number = table_column.IntegerField()
     """