+++ src/simmate/database/base_data_types/thermodynamics.py
@@ -20,6 +20,15 @@ class Thermodynamics(DatabaseTable):
 
     base_info = ["energy"]
 
+    api_filters = dict(
+        energy=["range"],
+        energy_per_atom=["range"],
+        energy_above_hull=["range"],
+        is_stable=["exact"],
+        formation_energy=["range"],
+        formation_energy_per_atom=["range"],
+    )
+
     energy = table_column.FloatField(blank=True, null=True)
     """
     The calculated total energy. Units are in eV.